Added env settings to backend server

+++ product/signals.py
@@ -98,6 +98,7 @@ def delete_image(sender, instance, **kwargs):
     s = sender.objects.get(pk=instance.pk)
 
     if (not s.item_image or s.item_image is not None) and (not s.item_image_webp or s.item_image_webp is not None):
-        s.item_image.delete(False)
-        s.item_image_webp.delete(False)
-        delete_dir(s.item_image.url.replace(s.item_image.filename, "").replace(s.item_image.format, "").replace(".", ""))
+        if item_image:
+            s.item_image.delete(False)
+            s.item_image_webp.delete(False)
+            delete_dir(s.item_image.url.replace(s.item_image.filename, "").replace(s.item_image.format, "").replace(".", ""))
